test_deploy/requirements.txt
# --- Web & API ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# --- HTTP / IO ---
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1

# --- Repo / AWS SDK ---
GitPython==3.1.40
boto3==1.34.0
botocore==1.34.0

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2
ipaddress==1.0.23

# --- Data / Analytics ---
numpy==1.24.3
pandas==2.1.4

# --- DB / ORM / Migrations ---
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# --- Cache / Queue ---
redis==5.0.8
celery==5.3.4
kombu==5.3.4

# --- Payments ---
stripe==8.0.0

# --- Infra / misc ---
requests==2.31.0
